start on chart
