+++ nutrinova-ui/package.json
@@ -14,6 +14,7 @@
     "@mui/icons-material": "^5.14.19",
     "@mui/lab": "^5.0.0-alpha.154",
     "@mui/material": "^5.14.11",
+    "@mui/x-charts": "^6.19.5",
     "@mui/x-data-grid": "^6.16.3",
     "@mui/x-date-pickers": "^6.18.3",
     "@tanstack/react-query": "^5.7.0",
